% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioDataScience-package.R
\docType{package}
\name{BioDataScience2018-package}
\alias{BioDataScience2018-package}
\title{A Series of Learnr Documents for BioDataScience (2018 version)}
\description{
Material for the course.
}
\section{Important functions}{

\itemize{
\item \code{\link[=run]{run()}} is the main function to launch learnr interactive documents
}
}

